integration_with_pymongo/requirements.txt
pymongo==3.12.0
python-dotenv==0.19.1